Fix websocket reconnection
