+++ js/session.js
@@ -1,14 +1,15 @@
 
+export let sessionId = null;
 export let websocket = null;
 
-export async function getSessionId({port, baseUrl, game, setSessionId}) {
+export async function fetchSessionId({port, baseUrl, game}) {
     let loaded = false;
     for (let i = 0; i < 20; i++) {
         try {
             const sessionRes = await fetch(`http://${location.hostname}:${port}/api/session`, {
                 method: "POST"
             });
-            setSessionId(await sessionRes.text());
+            sessionId = await sessionRes.text();
             const dataRes = await fetch(`${baseUrl}/api/load`);
             const dataText = await dataRes.text();
             game.deserialize(dataText);
@@ -20,7 +21,7 @@ export async function getSessionId({port, baseUrl, game, setSessionId}) {
     }
 }
 
-export function reconnectWebSocket(){
+export function reconnectWebSocket({port, game, onupdate = () => {}}){
     if(sessionId){
         websocket = new WebSocket(`ws://${location.hostname}:${port}/ws/${sessionId}`);
         websocket.binaryType = "arraybuffer";
@@ -28,9 +29,8 @@ export function reconnectWebSocket(){
             if (event.data instanceof ArrayBuffer) {
                 const byteArray = new Uint8Array(event.data);
                 game.deserialize_bin(byteArray);
-                postChunksDigest();
-                heartbeatOpacity = 1;
-                updateHeartbeatOpacity();
+                postChunksDigest(game);
+                onupdate();
             }
             else {
             // console.log(`Event through WebSocket: ${event.data}`);
@@ -39,8 +39,7 @@ export function reconnectWebSocket(){
                     if(game){
                         game.deserialize(data.payload);
                         postChunksDigest();
-                        heartbeatOpacity = 1;
-                        updateHeartbeatOpacity();
+                        onupdate();
                     }
                     // const payload = data.payload;
                     // const body = CelestialBody.celestialBodies.get(payload.bodyState.name);
@@ -50,6 +49,13 @@ export function reconnectWebSocket(){
                 }
             }
         });
-        websocket.addEventListener("open", postChunksDigest);
+        websocket.addEventListener("open", () => postChunksDigest(game));
     }
+}
+
+function postChunksDigest(game) {
+    game.uniformify_tiles();
+    const chunksDigest = game.serialize_chunks_digest();
+    // websocket.send(JSON.stringify({type: "ChunksDigest", payload: {chunks_digest: chunksDigest}}));
+    websocket.send(chunksDigest);
 }
